docs: added document for testhelper
