+++ docs/godoc.md
@@ -1,0 +1,6 @@
+# How to preview document on local machine
+
+1. Install `godoc` by running command `go install golang.org/x/tools/cmd/godoc@latest`
+2. Go to project root folder, and run `godoc`
+3. Open `http://localhost:6060` in your browser
+4. Search package you needed (i.e. "testhelper", there are more than one result, make sure you pick the right one)
