+++ testhelper/csv_file_iterator_test.go
@@ -1,0 +1,12 @@
+package testhelper
+
+
+func ExampleCsvFileIterator() {
+   iter := NewCsvFileIterator("/path/to/foobar.csv")
+   defer iter.Close()
+   for iter.HasNext() {
+	 row := iter.Fetch()
+	 println(row["name"])   // foobar
+	 println(row["json"])   // {"url": "https://example.com"}
+   }
+}
